+++ NEWS.md
@@ -4,10 +4,11 @@
 
 #### Others
 
+* updated news.md and description
 * some revisions to file path handling to avoid warnings
 * updated news.md
 
-Full set of changes: [`v0.2.3...6b57449`](https://gitlab.miracum.org/miracum/dqa/dqastats/compare/v0.2.3...6b57449)
+Full set of changes: [`v0.2.3...f6b315f`](https://gitlab.miracum.org/miracum/dqa/dqastats/compare/v0.2.3...f6b315f)
 
 ## v0.2.3 (2022-02-04)