+++ NEWS.md
@@ -1,14 +1,15 @@
 # DQAstats NEWS
 
-## Unreleased (2022-02-08)
+## Unreleased (2022-02-11)
 
 #### Others
 
+* updated news.md, adaptions to new diztools/dizutils
 * updated news.md and description
 * some revisions to file path handling to avoid warnings
 * updated news.md
 
-Full set of changes: [`v0.2.3...f6b315f`](https://gitlab.miracum.org/miracum/dqa/dqastats/compare/v0.2.3...f6b315f)
+Full set of changes: [`v0.2.3...c038c44`](https://gitlab.miracum.org/miracum/dqa/dqastats/compare/v0.2.3...c038c44)
 
 ## v0.2.3 (2022-02-04)